Alter the behavior of `gzochid_application_transactional_task_worker' so that it accepts a timeout argument (which may be `NULL') that can be passed to `gzochid_transaction_execute_timed'.
This is not the greatest patch in the world. In particular, the use of
gpointer[2] could probably be improved upon - by, say, a custom type that
encapsulates the input to the worker function, since it is part of the
internal "public" API. Furthermore, it's not currently possible to free
heap-allocated instances of this array once they are no longer used - no
worse than some other memory leaks in the server, but no better, either.

* gzochi-server/src/app.c (gzochid_application_transactional_task_worker):
  Change the semantics of the `data' argument such that it is expected to
  be a two-element array giving an optional timeout, followed by the
  data argument to `transactional_task_worker'.
  (gzochid_application_client_logged_in,
  gzochid_application_client_disconnected,
  gzochid_application_session_received_message) Pass the default
  transaction timeout.
* gzochi-server/src/data.c (gzochid_data_persistence_task_new,
  gzochid_data_prefix_binding_persistence_task_new): Pass the default
  transaction timeout.
  (gzochid_data_persistence_task_free): Free the gpointer array.