Fix a race condition in session removal during disconnect
The disconnect callback task and the session cleanup task were getting
called at the same time, and the cleanup task was usually winning.
To make this more deterministic, use the resubmitting task execution
system: the callback task invokes the cleanup task's payload when it's done with the application's business; if the callback task fails enough,
the cleanup task will be executed as the cleanup task for the task
execution. Tricky.

* gzochi-server/src/app.c (gzochid_application_client_disconnected):
  Build a single transactional task that manages the resubmission and
  cleanup cycle for the callback and cleanup tasks.
* gzochi-server/src/scheme.c
  (gzochid_scheme_application_disconnected_worker): Add some more
  detailed logging for error conditions. Invoke
  `gzochid_client_session_disconnected_worker' to remove the session on
  all control flows.
* gzochi-server/src/session.c (remove_session): Remove the session
object
  even if the Scheme-side session has already been removed.