Refactor `gzochid' build to use Libtool "convenience libraries."
Now that there are multiple binaries that are derived from the same
universe of source files, it makes sense to link these binaries against a
single large static library that includes all the symbols they might need.
To slim down the resulting binaries, pass `-ffunction-sections' to CC; pass
`--gc-sections' to ld; and/or use strip.

* gzochi-server/src/Makefile.am (noinst_LTLIBRARIES, libgzochid_la_CFLAGS,
  libgzochid_la_SOURCES, libgzochid_la_LIBADD): New sections.
  (gzochid_SOURCES): Reduce to just `gzochid.c'.
  (gzochid_LDADD, gzochi_dump_LDADD, gzochi_load_LDADD): Reduce to just
  `libgzochid.la'.
  (gzochi_dump_SOURCES): Reduce to just `gzochi-dump.c' and `toollib.c'.
  (gzochi_load_SOURCES): Reduce to just `gzochi-load.c' and `toollib.c'.
  Remove all explicit .c -> .h dependencies.
* gzochi-server/src/api/Makefile.am (noint_LIBRARIES): Convert to
  `noinst_LTLIBRARIES'.
  (libgzochid_server_api_a_SOURCES): Convert to
  `libgzochid_api_la_SOURCES'.
  Remove all explicit .c -> .h dependencies.