Fix long-standing issue with console corruption in AberMUD example game.
(N)curses is not thread-safe out of the box, and it's not easy to make it
so. The provided mutex-callback functions (e.g., `use_window') don't seem
to work with blocking calls that may perturb state (e.g., `wgetch').
Instead, tighten up the client's use of the API such that it all happens
from a single thread via a primitive "dispatch" mechanism using locks and
monitors.

* examples/abermud/client/src/console-curses.c
  (abermud_console_event_type): New enum.
  (console_event_lock, console_event_cond, console_event_type,
  console_event_text): New static globals.
  (abermud_console_init): Set halfdelay = 1.
  (append_output, set_prompt, set_title, clear_console,
  dispatch_console_event): New functions.
  (abermud_console_append_output, abermud_console_set_prompt,
  abermud_console_set_title): Rewrite such that they queue a console update
  event instead of doing the update themselves.
  (abermud_console_getch): Rewrite with a loop that polls for input and
  dispatches any pending console events.