bug #38136: Conditions raised from debugging context lose their stack context
This was happening because the transaction execution code was trapping the
exception and saving it to a variable. Re-throwing the contents of that
variable created a new prompt / stack which was less useful. The fix allows
the original exception to reach the remote REPL prompt while still cleaning
up the transaction in a dynamic unwind handler.

* gzochi-server/src/api/admin.c (with_application_unwind_handler): New
  function.
  (primitive_with_application): Rewrite transaction management around the
  specified thunk in terms of `gzochid_transaction_begin' and
  `gzochid_transaction_end'.