Significant refactoring to decouple modules and facilitate testing.
The `app' module has been split out into the additional `app-task',
`callback', and `lifecycle' modules; the `scheme' modules has been split
out in to the additional `scheme-task' module. This change should make it
easier for unit tests to target smaller bits of functionality and reduce
the number of modules they need to build against. It should be possible
for tests that only need to construct application contexts to do so without
linking against the transaction subsystem. Among other things, this commit
removes the need for many of the "mock" files in the tests directory,
which were challenging to keep synchronized and could not be easily shared
to subdirectories.

[Complete changeset truncated...]