Fix sloppy lock handling on rollback in gzochid data participant.
The data transaction participant locks the mutex for the `oids' and `names'
stores during prepare and unlocks them on rollback or commit. But since
rollback can be called without prepare, this means that an unlocked mutex
can sometimes be double-unlocked. Interestingly, this manifests as a test
failure trying to double-unlock during commit, but once a double-unlock
happens, subsequent behavior is undefined, so it's not really that strange.

* gzochi-server/src/data.c, data.h (gzochid_data_transaction_context): Move
  struct definition to `data.c', since it's not a public part of the data
  participant.
* gzochi-server/src/data.c (gzochid_data_transaction_context): New struct
  field `needs_unlock'.
  (data_prepare): Set `needs_unlock' if prepare is called.
  (data_rollback): Only unlock the store mutexes if they were successfully
  locked.