Fix two bugs in the gzochid memory store related to deletion.
The first issue has to do with the check performed during delete that
asserts that the node has no modifications to its key or value. The
intra-transaction deletion code clears value modifications on delete but
not key modifications.

The second issue is the use of `unlink_node' in `apply_modification', which
is not safe because it attempts to dereference sibling nodes via the
effective header, which changes during the commit process. Nothing that
could possibly attempt to acquire a new lock should be done during commit
(or rollback).

* gzochi-server/src/storage-mem.c (mark_node_deleted): Clear the `new_key'
  if the datum is non-NULL.
  (apply_modification): Remove call to `unlink_node', which is already
  called in the body of the transaction inside `mark_node_deleted'.