Fix a race on channel messaging with cleanup of disconnected clients.
The channel service was grabbing a reference to a client from the oid
mapping table and using it to schedule messaging tasks, but the client
could sometimes disconnect and be freed before its messaging task was
executed. Instead, the service should schedule the tasks using the client's
session oid and have the tasks hold the oid mapping lock while looking up
the client and enqueueing the message.

* gzochi-server/src/channel.c (gzochid_channel_message): New struct fields
  `app_context' and `oid_str'. Remove struct field `client'.
  (gzochid_channel_message_new, gzochid_channel_payload_message_new): Bind
  `app_context' and `oid_str', not `client'.
  (commit_channel_side_effect_message): Hold the `client_mapping_lock'
  while looking up a client by oid and sending a message to it.
  (close_channel, send_channel_message, join_channel, leave_channel): Don't
  look up the client before scheduling the task.