Fix a critical issue related to concurrent channel message delivery.
The gzochid channel API uses the `SCM_BYTEVECTOR_CONTENTS' macro to access
the raw contents of the message bytevector, and the downstream channel
message delivery code used that buffer as-is. If the buffer was re-used by
Guile before the channel transaction commit and subsequent socket send, the
contents of the message could be overwritten. The solution is to
defensively copy the channel message within the scope of the
`primitive-send-channel-message' invocation.

* gzochi-server/src/api/channel.c: General code formatting and cleanup.
* gzochi-server/src/channel.c (gzochid_channel_payload_message_new,
  create_send_operation): Make defensive copies of message payload.
  (free_message, free_operation): New functions.
  (cleanup_side_effects_transaction): Use `free_message' during transaction
  cleanup.
  (close_channel, send_channel_message, join_channel, leave_channel): Free
  the channel operation structure once the operation has been processed.