bug #45745: Non-retryable condition in disconnect handler causes abort
Specifically, on sweeping expired client sessions on startup. This is
because that code was assuming that the disconnect transaction couldn't
possibly fail if given infinite execution time and no contention with other
threads, but it can totally still fail - because of an application error,
for example, in which case the application should merely fail to launch,
not cause a server assertion failure.

* gzochi-server/src/lifecycle.c (initialize_async_transactional): Check for
  a failure to disconnect an old session, and halt the application if
  detected.
* gzochi-server/src/session.c, session.h (gzochid_sweep_client_sessions):
  Modify prototype to take a `GError' and not both with a
  `gzochid_auth_identity'.
* gzochi-server/src/session.c (gzochid_session_error_quark): New function.
* gzochi-server/src/session.h (GZOCHID_SESSION_ERROR): New preprocessor
  macro.
  (GzochidSessionError): New enum typedef.
* gzochi-server/tests/Makefile.am (test_programs): Add `test-session'.
  (test_session_CFLAGS, test_session_SOURCES, test_session_LDADD): New
  declarations.
* gzochi-server/tests/test-session.c: New file.