Further stabilization work for periodic task cancellation test.
Account for the fact that setting a thread pool's size to 0 has no effect
on its willingness to take additional tasks from its queue. Apparently it
must be freed to fully halt. Also account for the fact that the number of
re-executions of a task that occur _after_ it has been canceled is pretty
hard to control. So don't bother.

* gzochi-server/tests/test-durable-task.c (test_periodic_cancel): Free the
  pool during teardown after stopping the queue thread. Cancel the task
  once it's executed _at least_ three times.