GZOCHI_COMMON_* take iii: Move abs_top_* references back to configure.ac.
The Makefile.am-based approach didn't work because it relied on the
variable being defined and exported by the top-level Makefile. ...Which
works fine for a full recursive build but not so much for, say, a one-off
`make check' in a tests/ subfolder.

This time, let's try setting the value of the variable that we `AC_SUBST'
to include references to stuff that's guaranteed to be available in the
Makefile environment.

* gzochi-glib-client/Makefile.am, gzochi-server/Makefile.am
  (GZOCHI_COMMON_CFLAGS, GZOCHI_COMMON_LIBS): Remove declarations.
* gzochi-glib-client/configure.ac, gzochi-server/configure.ac
  (GZOCHI_COMMON_CFLAGS, GZOCHI_COMMON_LIBS): New `AC_SUBST' declarations.
* gzochi-glib-client/src/Makefile.am, tests/Makefile.am,
  gzochi-server/src/Makefile.am, tests/Makefile.am, tests/api/Makefile.am,
  tests/auth/Makefile.am: `$(GZOCHI_COMMON_CFLAGS)' ->
  `@GZOCHI_COMMON_CFLAGS@'; `$(GZOCHI_COMMON_LIBS)' ->
  `@GZOCHI_COMMON_LIBS@'.