Significant refactoring of the embedded HTTP service context in gzochid.
Better separates the concerns of bootstrapping the server from mapping and
handling requests, making the service context more reusable (in a
metaserver, for example) and easier to test. The HTTP service context has
been rewritten to provide a minimal HTTP server framework that should make
it easier to add and test new handler code as well.

* gzochi-server/src/Makefile.am (noinst_HEADERS): Add `httpd-app.h'.
  (libgzochid_la_SOURCES): Add `httpd-app.c'.
* gzochi-server/src/admin.c (initialize): Register the handlers.
* gzochi-server/src/httpd-app.c, httpd-app.h: New files.
* gzochi-server/src/httpd.c, httpd.h: Rewrite to implement HTTP server
  framework.
* gzochi-server/tests/Makefile.am (test_programs): Add `test-httpd'.
  (test_httpd_CFLAGS, test_httpd_SOURCES, test_httpd_LDADD): New
  declarations.
* gzochi-server/tests/test-httpd.c: New file.