Clarify the ownership of the fields in the change array in a changeset.
When it's deserialized from a socket, it should be owned by the changeset
object itself and freed when that object is freed. When it's constructed in
client code or test code, however, the change objects should be owned by
the caller. To that end, this patch makes the ownership configurable by
allowing callers to specify a "free function" for change fields.

* gzochi-server/src/data-protocol.c, data-protocol.h
  (gzochid_data_changeset_new_with_free_func): New function.
* gzochi-server/src/data-protocol.c (gzochid_data_changeset_new): Delegate
  to `gzochid_data_changeset_new_with_free_func'.
  (gzochid_data_changeset_free): Invoke the free function if available.
  (gzochid_data_protocol_changeset_read): Construct the changeet with
  `change_clear' as the free function.
* gzochi-server/src/data-protocol.h (gzochid_data_changeset): New struct
  field `free_func'.
* gzochi-server/src/tests/test-data-protocol.c (test_changeset): Clean up
  change fields.
* gzochi-server/src/tests/test-dataserver.c (test_process_changeset): Clean
  up change fields.