Refactor channels implementation to avoid interactions with data store.
The membership of a channel, at least with respect to local connections to
a single application server node, is not something that needs durable
persistence, since upon container restart, those connections will all be
gone. And maintaining it durably creates all sorts of contention. This
patch makes channel membership ephemeral.

* gzochi-server/src/app.c (gzochid_application_context_new): Initialize
  `channel_oids_to_local_session_oids' and `channel_mapping_lock'.
  (gzochid_application_context_free): Free
  `channel_oids_to_local_session_oids' and `channel_mapping_lock'.
* gzochi-server/src/app.h (gzochid_application_context): New struct fields
  `channel_oids_to_local_session_oids' and `channel_mapping_lock'.
* gzochi-server/src/channel.c: General rewrite to remove dependency on
  data store management of session membership.