Store responses from the dataserver in a cache in front of the mem store.
It turns out that writing the response from the dataserver directly into
the mem store is a bit fraught, because it can only be done
transactionally, and any transaction that inserts it will likely disrupt
transactions waiting for it. Instead, put it into a hash table to be
checked as a fallback by any waiting read transactions.

* gzochi-server/src/storage-dataclient.c (dataclient_database): New struct
  field `value_cache'.
  (dataclient_transaction): New struct field `deleted_keys'.
  (lock_success_callback): Add value to cache instead of mem store.
  (open): Initialize the value cache.
  (close_store): Tear down the value cache.
  (cleanup_transaction): Cleanup deleted keys list.
  (transaction_commit): Purge deleted keys from value cache.
  (callback_data_compare, is_deleted, get_from_cache): New functions.
  (transaction_get, transaction_get_for_update): Consult the value cache
  upon a "miss" against the mem store.
  (transaction_delete): Append target key to deleted keys list.