Explicit lifecycle management for application context fields.
Until bug #48768 is fixed, calling `gzochid_application_context_free' isn't
safe to call if the context mutex object hasn't been initialized via a call
to `gzochid_context_init'... or if it's explicitly initialized in a kinda
ugly way as in the following changes.

This seems to be a particular problem for versions / builds of GLib in
which `GMutex' is implemented in terms of `pthread_mutex_t'.

* gzochi-server/src/gzochi-migrate.c, tests/api/test-tx.c, test-app-task.c,
  test-data.c, test-game-protocol.c, test-scheme-task.c, test-scheme.c,
  test-session.c: Initialize various context mutexes so that they can be
  safely torn down by the "free" functions.