Remove flaky test for durable task cancellation.
This test has always been a bit problematic, and especially prone to
deadlock recently given its reliance on both explicit locking (via GMutex)
and implicit locking (via access to the storage interface) in order to run
tasks and protect access to the counter variable. The way to fix it for
real is to remove all explicit locking (as you would do when writing a real
gzochi game application) but that's gonna take a re-think of the test logic
as well.

* gzochi-server/tests/test-durable-task.c (test_worker_counter,
  counter_worker_serialization, test_worker_counter_serialization,
  task_serialization): Removed fields.
  (test_counter_worker, test_counter_worker_serializer,
  test_counter_worker_deserializer, serialize_test_worker_counter,
  deserialize_test_worker_counter, finalize_test_worker_counter,
  test_periodic_cancel_inner0, test_periodic_cancel_inner1,
  test_periodic_cancel): Removed functions.
  (main): Remove test registration for `/durable-task/periodic/cancel'.