Fix broken assumption about lock release / callback data in data client.
The assumption was that if a lock is requested and then upgraded (from read
to write) the release callback will only be invoked once. So it makes sense
to clean up the callback data on the success callback of an upgrade, since
that's the last opportunity to do so. ...Except the assumption is wrong:
The release callback will happen as many times as the success callback is
invoked; it's clever enough to ignore locks that have already been cleaned
up by earlier releases. So don't bother being clever with the callback
cleanup.

* gzochi-server/src/storage-dataclient.c (lock_success_callback): Don't
  free the callback data.
* gzochi-server/tests/test-storage-dataclient.c
  (dataclient_release_closure): New struct field `qualified_key'.
  (GzochidDataClient): Change type of struct field `release_closures' to
  `GList'.
  (execute_release_closure): Free the qualified key.
  (gzochid_data_client_init): Initialize the release closures list.
  (create_release_closure): Accept qualified key argument.
  (process_response_async): Append release closure to list.
  (find_release_closure): New function.
  (release_key, release_key_range): Search the release closure list by key.
  (dataclient_storage_fixture_teardown): Clean up the closure list.