Merging a changeset on data server should not affect lock state.
Important to use `check' rather than `check-and-set' during changeset
processing, since it could happen that a node submits a changeset for which
it does not hold all the write locks but for which write locks are
available. If this happens, the data server could mark those locks acquired
on behalf of the node - but the node wouldn't know that it had acquired
them and thus would never release them.

* gzochi-server/src/dataserver.c
  (gzochi_metad_dataserver_process_changeset):
  `gzochid_lock_check_and_set' -> `gzochid_lock_check'.