Refactor the dataclient storage engine a bit to ease fix for cache bug.
See bug #50647

The core of this refactoring "lifts" the management of point locks and lock
requests, as well as the "value cache" up from the `dataclient_database'
struct to the owning `dataclient_environment' in order to ease some
forthcoming changes to mutex management during lock release. A new mutex
for lock table management is added to the environment; range lock
management continues to live at the database level, as it is unaffected by
the cache consistency bug.

* gzochi-server/src/storage-dataclient.c: Too many changes to describe.