bug #50646: valgrind warning during mem store delete: Use of uninitialised value of size 8
Interesting race condition: Two thread preparing to lock a node for
deletion interleave their requests for the node's effective page for
writes. This block isn't synchronized, and the result is that a new scratch
page is allocated twice, potentially wiping out one thread's work with an
unitialized block. The fix is to make sure that all transactions seize the
appropriate lock before asking for the effective page.

* gzochi-server/src/storage-mem.c (transaction_delete): Ensure write lock
  before requesting effective page.