Update `gzochi-guile-client' for Guile 2.2 compatibility.
This change is a backwards-compatible fix for some mutex sloppiness: The
code was relying on SRFI-18 mutexes (a) being compatible with Guile native
mutexes; in 2.2, that's no longer the case, and (b) being recursive,
contrary to spec; Guile 2.2 fixes that for SRFI-18. The fix is to ditch the
use of SRFI-18, since Guile native mutexes provide everything we need.

* gzochi-guile-client/src/gzochi/client.scm: Remove import for
  `(srfi :18)'.
* gzochi-guile-client/src/gzochi/main-loop.scm: Remove import for
  `(srfi :18)'.
  (gzochi:main-loop): Specify 'recursive in flags argument to `make-mutex'.