bug #50647: In distributed mode, the dataclient storage engine gives inconsistent responses
The root cause of this issue was that the delegate store was not getting
cleared when a new a value arrived from the meta server, and that new value
would sometimes get ignored. Clearing the delegate store non-disruptively
is not straightforward, so this patch works around the issue by tagging
values with their timestamps and preferring more recent values. A
subsequent patch will ensure that the delegate store is properly "garbage
collected."

* gzochi-server/src/storage-dataclient.c (lock_unref): Don't lock the
  environment; expect callers to do it.
  (prefix_with_timestamp, prefix_bytes_with_timestamp,
  create_empty_timestamped_value, create_empty_timestamped_bytes,
  get_prefixed_cache_value, get_value_from_cache, get_timestamp_from_cache,
  get_internal): New functions.
  (lock_success_callback): Cache empty values.
  (cleanup_transaction): Lock the environment to call `lock_unref'.
  (transaction_get, transaction_get_for_update): Rewrite in terms of
  `get_internal'.
  (transaction_put): Write prefixed values to delegate store.
* gzochi-server/tests/test-storage-dataclient.c
  (test_cache_store_consistency): New function.
  (main): Bind `/storage-dataclient/consistency/cache-store'.